requirements.txt
pandas
pyyaml
pyarrow
networkx
matplotlib
streamlit